Código arrumado, opção de editar produto e estilo tailwind

+++ src/main/java/com/example/PharmaPhorm/produto/ProdutoController.java
@@ -34,6 +34,9 @@ public class ProdutoController {
 
     @DeleteMapping("/produtos/{id}")
     public void delete(@PathVariable Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Produto não encontrado");
+        }
         repository.deleteById(id);
     }
